@@ -5,23 +5,25 @@
   "main": "index.js",
   "author": "Mark Njung'e",
   "license": "MIT",
+  "type": "module",
   "scripts": {
-    "lint": "eslint \"**/*.{ts,js}\"",
-    "lint:fix": "eslint \"**/*.{ts,js}\" --fix"
+    "lint": "eslint",
+    "lint:fix": "eslint --fix"
   },
   "repository": {
     "type": "git",
     "url": "https://github.com/MarkNjunge/eslint-config-ts"
   },
   "peerDependencies": {
-    "eslint": ">= 8",
-    "@stylistic/eslint-plugin": ">= 1.5.4"
+    "eslint": ">= 9"
   },
   "devDependencies": {
-    "@stylistic/eslint-plugin": "^1.5.4",
-    "@typescript-eslint/eslint-plugin": "^6.4.0",
-    "@typescript-eslint/parser": "^6.4.0",
-    "eslint": "^8.47.0",
-    "typescript": "^5.1.6"
+    "@eslint/js": "^9.25.1",
+    "@stylistic/eslint-plugin-js": "^4.2.0",
+    "@stylistic/eslint-plugin-ts": "^4.2.0",
+    "eslint": "^9.25.1",
+    "globals": "^16.0.0",
+    "typescript": "^5.8.3",
+    "typescript-eslint": "^8.31.0"
   }
 }
